Add files via upload
This program, autoOff, is a practice to use python to manipulate registry and windows command.
Every time the system boot, it will be shut in one minute.
At the beginning, I tried to use C++ to finish this executalbe. But the path, environment and pack problems always confuse me. So Python become my first choice.