Add files via upload
Hi
This is a practice of x32 dbg. I download this exe file from https://crackmes.one/
The red circle in the png file is the solution of this question. But the answer is encrypted by MD5.
The c93ccd78b2076528346216b3b2f701e6 is the code and the decrypted answer is admin1234