Add files via upload
Hi
This practice is analyzed by IDA.
The answer to this practice is rXv3rs1nX_f7W! Note that the X in there means any character or number.
The advanced_static_structure.png contains a green circle that is used to check the input characters. The green 0 to 6 corresponds to the function check0 to check6. Also, check0 checks the index 0 and 7 of the input, check1 checks the index 1 and 8, and so on.
Note that some of these functions directly do the compare job to check the input, some are not. For those not, I analyzed them from their outputs to see what they need.