Add files via upload
Hi
This is a port scanner. It automatically gets the local IP (Depends on different computers. In my laptop the default network interface is eth0, but you can change in the NETWORK.) and netmask to generate all possible IPs.
It will do scan the port from 1 to 100 (If you like others you can change in the LOW_PORT and HIGH_PORT).
It will print the IPs which are used currently and the MAC address related. Also, the open port will be shown.