Add files via upload
Hi
This is a practice of upload file attack.
It is based on OWASP and KALI systems.
In the upload file attack. Set a proxy in firefox on KALI and use the burpsuit to do the content change and forward package.
In my practice, I set the security as high and upload a PHP script.
In the proxy_change_package image, as you can see, I change the cookie from high to low and content-type from applicaiton/x-php to image/jpeg.
The get_shell image shows the result of the pwd command.